% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{all.equalcn}
\alias{all.equalcn}
\title{checks if the colnames of two dataframes are.equal.}
\usage{
\method{all}{equalcn}(x, y)
}
\arguments{
\item{x}{a dataframe}

\item{y}{a dataframe}
}
\description{
checks if the colnames of two dataframes are.equal.
}
