% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charindex.R
\name{charindex}
\alias{charindex}
\title{Index numbers for all letters in a string}
\usage{
charindex(char)
}
\arguments{
\item{char}{a character vector}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
index numbers for all letters in a string
}
\examples{
\dontrun{
charindex(c('Jens Peter Hanson', 'dato 2020-01-01'))
}
}
