% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{closes all open data viewers in linux}
\usage{
cv()
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
requires wmctrl to be installed. sudo apt-get install wmctrl in Ubuntu (and perhaps also other debian-based systems)
}
\examples{
\dontrun{
function(arg1)
}
}
