% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_copy.R
\name{dir_copy}
\alias{dir_copy}
\title{copy directoy}
\usage{
dir_copy(from, to)
}
\arguments{
\item{from}{from which folder}

\item{to}{to which folder}
}
\value{
This function copies a directory to another location.
}
\description{
Copy a directory. Max / Linux only for now
}
\examples{
\dontrun{
dir_copy('/from/here/', '/to/here/')
}
}
