% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_table.R
\name{dir_table}
\alias{dir_table}
\title{common file-operations-function}
\usage{
dir_table(dir = "", ...)
}
\arguments{
\item{dir}{directory to data.tablize}
}
\value{
This function returns a \code{data.table} with file-information.
}
\description{
creates a data.table with enhanced file-information from a directory. additional arguments are passed to dir_info from the fs-library.
}
\examples{
\dontrun{
dir_table('/path/to/folder')
}
}
