% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_write_clip.r
\name{dt_write_clip}
\alias{dt_write_clip}
\title{short description 1}
\usage{
dt_write_clip(df1, var, n = 10, all = FALSE, sample = TRUE)
}
\arguments{
\item{df}{a dataframe}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
