% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcolc.R
\name{dtcolc}
\alias{dtcolc}
\title{short description 1}
\usage{
dtcolc(dt, ...)
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
