% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtquantile.R
\name{dtquantile}
\alias{dtquantile}
\title{short description 1}
\usage{
dtquantile(x, start = 0, slut = 1, interval = 0.05, char = TRUE, na.rm = TRUE)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
