% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttools.R
\docType{package}
\name{dttools}
\alias{dttools}
\alias{dttools-package}
\title{dttools: useful tools built around data.table}
\description{
custom tools for wicked kids
}
\author{
\strong{Maintainer}: Emil Begtrup-Bright \email{emil.bb@gmail.com} (\href{https://orcid.org/0000-0002-5475-4057}{ORCID})

}
