% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttools_desc.R
\name{dttools_desc}
\alias{dttools_desc}
\title{short description 1}
\usage{
dttools_desc(dt)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
