% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttools_na.R
\name{dttools_na}
\alias{dttools_na}
\title{short description 1}
\usage{
dttools_na(dt, var1, by = NA, numeric = FALSE)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
