% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{dups}
\alias{dups}
\title{find dups in different ways}
\usage{
dups(DT, by = colnames(DT), index = FALSE)
}
\arguments{
\item{DT}{a data.table with suspcted dups}

\item{by}{the columns by which the dups should be detected}

\item{index}{if TRUE, the indices are returned instead of the rows of the DT with dups.}
}
\value{
This function returns a \code{data.table} with all the dups by the specified columns, instead of 'all the dups minus one' as is standard in data.table
}
\description{
these functions adds functionality lacking in data.table and other packages that are concerned with dups
}
\details{
do different stuff, yes, that's what happens!
}
\examples{
dups(dupstestdata)
dups(dupstestdata, index = TRUE)
}
