% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupsN.R
\name{dupsN}
\alias{dupsN}
\title{find number of dups,}
\usage{
dupsN(DT, by = colnames(DT), mode = "all")
}
\arguments{
\item{DT}{A data.table}

\item{by}{the variables used to identify the dups}
}
\value{
This function returns an \code{integer} value with the number of duplicates
}
\description{
here is a long description
}
\details{
not sure about this right here
}
\examples{
\dontrun{
dupsN(DT, by='variable1')
}
}
