% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupstestdata.R
\docType{data}
\name{dupstestdata}
\alias{dupstestdata}
\title{Random Duplicated Data}
\format{
A data frame of 7 rows and 3 columns
\describe{
 \item{id}{just random ID}
 \item{var1}{random numbers, some of them dups}
}
}
\source{
Randomly generated data
}
\usage{
dupstestdata
}
\description{
A dataset containing randomly generated duplicated data.
}
\keyword{datasets}
