% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esave.R
\name{e_internal_format.object_size}
\alias{e_internal_format.object_size}
\title{short description 1}
\usage{
e_internal_format.object_size(
  x,
  units = "b",
  standard = "auto",
  digits = 1L,
  ...
)
}
\description{
this is just a copy of utils:::format.object_size(), which is an internal function to the base R utils package. You're not allowed to use internal functions from other packages, i.e. the "..." operator, at least not if you want your package on CRAN. To come around this, I just copied the function here.
}
