% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecolor_pal.R
\name{ecolor_pal}
\alias{ecolor_pal}
\title{short description 1}
\usage{
ecolor_pal(colorx, background = "light gray", size = 16)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
ecolor_pal(ecolor_data[grp \%in\% 'paul tol 14', hex])
${1:}
}
}
