% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecolor_pie.R
\name{ecolor_pie}
\alias{ecolor_pie}
\title{short description 1}
\usage{
ecolor_pie(col)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
ecolor_pie(ecolor_data[grp \%in\% 'paul tol 14', hex])

}
}
