% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{emean}
\alias{emean}
\title{mean-function with reasonable defaults}
\usage{
emean(x, f = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{a vector}

\item{f}{should the number be formatted for readability?}
}
\description{
mean-function with reasonable defaults
}
