% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_size_formated.R
\name{file_size_formated}
\alias{file_size_formated}
\title{file size, nicely formatted}
\usage{
file_size_formated(size)
}
\arguments{
\item{size}{size to format}
}
\value{
This function returns \code{a character vector} with the size appropriately formatted.
}
\description{
file size formatted smartly to use a unit appropiate for it's size
}
\examples{
\dontrun{
file_size_formated(1:10)
}
}
