% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_installer.R
\name{github_installer}
\alias{github_installer}
\title{installs a R package from github}
\usage{
github_installer(url = NULL, ...)
}
\arguments{
\item{url}{link to a github r package}
}
\value{
This function  \code{the url} that was given to it, after it has installed the package
}
\description{
installs a R package from github

installs a R package from github
}
\examples{
\dontrun{
github_installer('emilbebri/dttools')
}
}
