% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_at.R
\name{insert_at}
\alias{insert_at}
\title{insert elements in a vector at a specifik position}
\usage{
insert_at(vector1, pos, ...)
}
\arguments{
\item{vector1}{a vector}

\item{pos}{positions of were the elements/vectors should be inserted. Length must be the same as the number of vectors inserted.}

\item{...}{vectors that should be inserted}
}
\value{
This function creates, by reference, four variables with run- and spell-length-type information.
}
\description{
insert elements in a vector at a specifik position. can't be in the beginning of a vector, perhaps fix sometime #todo#
}
\examples{
\dontrun{
# first example
a1 = c(1:10)
b1 = c('SE MIG')
c1 = c('A', 'B')

#indsaetter b efter 3. element og d efter 7. element i a 
insert_at(a1, c(3,7), b1, c1) 
#andet eksempel
insert_at(a1, c(4,7,9), "hello", "this", "world") 

}
}
