% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{intersectcn}
\alias{intersectcn}
\title{checks if the intersection of colnames of two dataframes are.equal.}
\usage{
intersectcn(x, y)
}
\arguments{
\item{x}{a dataframe}

\item{y}{a dataframe}
}
\description{
checks if the intersection of colnames of two dataframes are.equal.
}
