% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage-.ls.objects.R
\name{memory_usage}
\alias{memory_usage}
\title{short description 1}
\usage{
memory_usage(..., n = 10)
}
\arguments{
\item{n}{top n - default is the 10 most memory consuming}

\item{all}{should all the objects be returned? Default is FALSE.}
}
\value{
This function returns a \code{data.table} with memory usage for the objects in the R session
}
\description{
descriptin 2??
}
\examples{
\dontrun{
memory_usage(n=15)
memory_usage(all=TRUE)
}
}
