% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{nr2}
\alias{nr2}
\title{nrows for one data.frame minus the nrows of another data.frame}
\usage{
nr2(dtx, dty, mode = "-")
}
\arguments{
\item{dtx}{A data.frame}

\item{dty}{A data.frame}
}
\value{
This function returns an \code{integer} value with the difference in number of rows between the two data.frames (dtx-dty)
}
\description{
that's about it
}
\details{
not sure about this right here
}
\examples{
\dontrun{
nr2(dtx, dtx)
}
}
