% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{nrf2}
\alias{nrf2}
\title{nrows for one data.frame minus the nrows of another data.frame, with readable formatting}
\usage{
nrf2(dtx, dty)
}
\arguments{
\item{dtx}{A data.frame}

\item{dty}{A data.frame}
}
\value{
This function returns an \code{character} vector with the difference in number of rows between the two data.frames (dtx-dty), with nicely readable formatting
}
\description{
that's about it
}
\details{
not sure about this right here
}
\examples{
\dontrun{
nr2(dtx, dtx)
}

}
