% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_long.R
\name{sample_long}
\alias{sample_long}
\title{sample from panel or longitutional data, where you'd like all the observations from a unique id (that is sampled)}
\usage{
sample_long(datx, idvar, n = 5)
}
\arguments{
\item{datx}{data.table or data.frame}

\item{idvar}{variable that identifies a unit across observations}

\item{n}{number of units (*not* number of observations)}
}
\value{
This function returns a \code{data.table} with samples from a larger data.table
}
\description{
not sure about this right here
}
\examples{
\dontrun{
function(arg1, arg2 etc)
}
}
