% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_flush_old.R
\name{script_flush_old}
\alias{script_flush_old}
\title{short description 1}
\usage{
script_flush_old(dir = "r/", extension = "r")
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
