% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_management_internal.R
\name{script_management_internal}
\alias{script_management_internal}
\title{short description 1}
\usage{
script_management_internal(dir = "r/", extension = "r")
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
