% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{setdiff2}
\alias{setdiff2}
\title{turn around x and y in the setdiff function}
\usage{
setdiff2(x, y, ...)
}
\arguments{
\item{x}{a vector}
}
\description{
turn around x and y in the setdiff function
}
