% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{setdiffcn}
\alias{setdiffcn}
\title{takes setdiff on colnames of two dataframes}
\usage{
setdiffcn(x, y)
}
\arguments{
\item{x}{a dataframe}

\item{y}{a dataframe}
}
\description{
takes setdiff on colnames of two dataframes
}
