% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{setdiffcn2}
\alias{setdiffcn2}
\title{takes setdiff on colnames of two dataframes, but with the y dataframe as the first argument (turns around setdiffcn)}
\usage{
setdiffcn2(x, y)
}
\arguments{
\item{x}{a dataframe}

\item{y}{a dataframe}
}
\description{
takes setdiff on colnames of two dataframes, but with the y dataframe as the first argument (turns around setdiffcn)
}
