% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setnames_clean.R
\name{setnames_clean}
\alias{setnames_clean}
\title{clean colnames}
\usage{
setnames_clean(x, ...)
}
\arguments{
\item{x}{a data.table with dirty names on it}
}
\value{
This function returns \code{a data.table} with clean names. all arguments passed to janitor::clean_names().
}
\description{
clean colnames, janitor style but by reference
}
\examples{
\dontrun{
function(arg1)
}
}
