% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-setscn.R
\name{sets}
\alias{sets}
\title{list of all the possible sets between two vectors}
\usage{
sets(x, y)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}
}
\value{
This function returns a \code{list} with the three types of sets: the \code{intersect} between two vectors, and the sets exclusive for the one or the other vector.
}
\description{
list of all the possible sets between two vectors
}
\examples{
\dontrun{
sets(arg1)
}
}
