% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-setscn.R
\name{setscn}
\alias{setscn}
\title{list of all the possible sets between two vectors of colnames}
\usage{
setscn(x, y)
}
\arguments{
\item{x}{the first data.frame}

\item{y}{the second data.frame}
}
\value{
This function returns a \code{list} with the three types of sets: the \code{intersect} between two vectors of colnames, and the sets exclusive for the one or the other vector of colnames.
}
\description{
list of all the possible sets between two vectors
}
\examples{
\dontrun{
sets(arg1)
}
}
