% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspell-setspell_selected.R
\name{setspell_selected}
\alias{setspell_selected}
\title{short description 1}
\usage{
setspell_selected(DT, v1, idvar = "pnr", value)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
