% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadopen.r
\name{spreadopen}
\alias{spreadopen}
\title{short description 1}
\usage{
spreadopen(x)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
