% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{unioncn}
\alias{unioncn}
\title{takes the union of colnames of two dataframes}
\usage{
unioncn(x, y)
}
\arguments{
\item{x}{a dataframe}

\item{y}{a dataframe}
}
\description{
takes the union of colnames of two dataframes
}
