% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convinience_functions.R
\name{uniqueA}
\alias{uniqueA}
\title{are all rows unique, TRUE/FALSE}
\usage{
uniqueA(DT, ...)
}
\arguments{
\item{DT}{a data.frame}
}
\value{
This function returns a \code{logical value} TRUE if all rows are unique, FALSE if they are not. ...  is passed on to uniqueN from data.table, where the columns to check uniqueness to can be supplied as well.
}
\description{
are all rows unique, TRUE/FALSE.
}
\details{
are all rows unique, TRUE/FALSE
}
