% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2.R
\name{v2}
\alias{v2}
\title{opens data in spreadsheet program}
\usage{
v2(data, filter = TRUE, table = TRUE, n = 9000, all = FALSE, sample = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{filter}{should there be the autofilter function enabled in the spreadsheet?}

\item{table}{as a table?}

\item{n}{number of rows}

\item{all}{should all rows be used? if TRUE, overrules the n argument.}

\item{sample}{should the n rows be sampled of taken from the top down?}
}
\value{
This function opens a \code{data.frame} in a spreadsheet viewer. if asssigned to an object, the data.frame as well as manual changes in the spreadsheet will be saved to that object.
}
\description{
opens data in spreadsheet program
}
\details{
opens data in spreadsheet program
}
\examples{
\dontrun{
dat2 <- vi(dupstestdata)
}
}
