% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colc.R
\name{colc}
\alias{colc}
\title{regex a vector or columnnames of data.frame and return only the matches with the regex}
\usage{
colc(DT, x = NA, not = NA, plus = NA, ignore.case = TRUE)
}
\arguments{
\item{DT}{a character vector. if a data.frame, uses the colnames}

\item{x}{a regex to search for in DT}

\item{not}{a regex that will exclude these from argument x. Note that this uses grepl, and since grepl is not vectorized, if several elements are supplied, they are stringed together with '|' in grepl.}

\item{plus}{a character vector with specific elements from the string that will be included}

\item{ignore.case}{should cases be ignored? Default is TRUE}
}
\value{
This function returns a \code{character vector} that matches the regex search pattern
}
\description{
regex a vector or columnnames of data.frame and return only the matches with the rege.
}
\details{
regex a vector or columnnames of data.frame and return only the matches with the regex
}
\examples{
colc(dupstestdata)
}
