% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{dups}
\alias{dups}
\title{find dups in different ways}
\usage{
dups(DT, by = colnames(DT), var = NULL)
}
\arguments{
\item{DT}{a data.table with suspcted dups}

\item{by}{the columns by which the dups should be detected}

\item{var}{should the output be a variabel that designates duplicates (TRUE) / unique(FALSE). If set, the function does NOT return an output, but modifices the data.table by reference and adds the variabel}
}
\value{
This function returns a \code{data.table} with all the dups by the specified columns, instead of 'all the dups minus one' as is standard in data.table. also sets the key to the "by" variables for code efficiency.
}
\description{
This functions adds functionality lacking in data.table and other packages that are concerned with dups


do different stuff, yes, that's what happens!
}
\examples{
\dontrun{
dups(dupstestdata)
}
}
