% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupsN.R
\name{dupsN}
\alias{dupsN}
\title{find number of dups,}
\usage{
dupsN(DT, by = colnames(DT), mode = "all")
}
\arguments{
\item{DT}{A data.table}

\item{by}{the variables used to identify the dups}

\item{mode}{"all" outpus the number of all the dups identified, while "unique" outputs only the number of unique duplicates}
}
\value{
This function returns an \code{integer} value with the number of duplicates in the
}
\description{
descriptin 2??

not sure about this right here
}
\examples{
\dontrun{
dupsN(DT, by='variable1')
}
}
