% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esave.R
\name{esave}
\alias{esave}
\title{short description 1}
\usage{
esave(
  DT1,
  filename = "",
  filepath = "./",
  csv = TRUE,
  xlsx = FALSE,
  fst = TRUE,
  qs_preset = "archive",
  fst_preset = 100
)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
