% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_usage-.ls.objects.R
\name{memory_usage}
\alias{memory_usage}
\title{short description 1}
\usage{
memory_usage(..., n = 10)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
