% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_flush_old.R
\name{script_flush_old}
\alias{script_flush_old}
\title{flush previous iterations of an R script}
\usage{
script_flush_old(dir = "r/", extension = "r")
}
\arguments{
\item{dir}{directory to loook for R scripts in}

\item{extension}{which file extension should be used}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
flush previous iterations of an R script
}
\examples{
\dontrun{
function(arg1)
}
}
