% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_newest.R
\name{script_newest}
\alias{script_newest}
\title{short description 1}
\usage{
script_newest(
  scriptname,
  dir = "r/",
  extension = "r",
  path = TRUE,
  source = TRUE
)
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
\dontrun{
function(arg1)
}
}
