% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets-setscn.R
\name{sets}
\alias{sets}
\title{short description 1}
\usage{
sets(x, y)
}
\arguments{
\item{argument1}{arg-description}
}
\value{
This function returns \code{the url} blah blah blah
}
\description{
descriptin 2??
}
\examples{
function(arg1)
}
