% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspell-setspell_selected.R
\name{setspell}
\alias{setspell}
\title{creates spells and run-lengths}
\usage{
setspell(DT, var1, grp, pos = "end")
}
\arguments{
\item{DT}{a data.table}

\item{var1}{variable to calcute run- and spell-lengths on.}

\item{grp}{grouping variables within-which to calculate on}

\item{pos}{where should the variables be located?}
}
\value{
This function creates, by reference, four variables with run- and spell-length-type information.
}
\description{
creates spells and run-lengths, by reference
}
\examples{
\dontrun{
function(arg1)
}
}
