% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspell-setspell_selected.R
\name{setspell}
\alias{setspell}
\title{creates spells and run-lengths}
\usage{
setspell(dtx, grp, var1, pos = "end", max = TRUE)
}
\arguments{
\item{dtx}{A data.table}

\item{grp}{Grouping variables within-which to calculate on}

\item{var1}{Variable to calcute run- and spell-lengths on.}

\item{pos}{Where should the variables be located in the data.table?}

\item{max}{Should the DT include the max run and spell lengths?}

\item{rename}{Should the spells and run-lengths be named after the original variable or generically named? (mostly used as TRUE with the setspell_selected variable)}
}
\value{
This function creates, by reference, four variables with run- and spell-length-type information.
}
\description{
creates spells and run-lengths, by reference
}
\examples{
\dontrun{
function(arg1)
}
}
