% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2.R
\name{v2}
\alias{v2}
\title{opens data in spreadsheet program}
\usage{
v2(
  data,
  filter = TRUE,
  table = TRUE,
  n = 9000,
  all = FALSE,
  sample = FALSE,
  engine = "writexl"
)
}
\arguments{
\item{data}{a data.frame}

\item{filter}{should there be the autofilter function enabled in the spreadsheet?}

\item{table}{as a table?}

\item{n}{number of rows}

\item{all}{should all rows be used? if TRUE, overrules the n argument.}

\item{sample}{should the n rows be sampled of taken from the top down?}

\item{engine}{which package should be used for convertion to xlsx? default is readxl. openxlsx has better setup of view options (such as using the excel-table function for autosorting), but in some rare cases it will fail on encoding of characters, deleting all information in the process, with no warning (that I can figure out). So it's not default for that reason, but will probably work fine most of the time. Just be careful about data loss when using it.}
}
\value{
This function opens a \code{data.frame} in a spreadsheet viewer. if asssigned to an object, the data.frame as well as manual changes in the spreadsheet will be saved to that object.
}
\description{
opens data in spreadsheet program
}
\details{
opens data in spreadsheet program
}
\examples{
\dontrun{
dat2 <- vi(  dupstestdata)
}
}
